warnings

+++ src/Step/Step6DockerProxy.php
@@ -11,10 +11,15 @@ class Step6DockerProxy extends Step0Base {
    * Run!
    */
   public function run() : bool {
-    $this->execCommands([
-      "cp -r {$_ENV['ASSETS']}/docker-proxy /opt/docker-proxy",
-      "/opt/docker-proxy/start.sh",
-    ]);
+    if (!file_exists('/opt/docker-proxy')) {
+      $this->execCommands([
+        "cp -r {$_ENV['ASSETS']}/docker-proxy /opt/docker-proxy",
+        "/opt/docker-proxy/start.sh",
+      ]);
+    }
+    else {
+      $this->io->warning("docker-proxy already exists");
+    }
     return 1;
   }
